requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
Werkzeug==2.3.7

firebase-admin==6.5.0
google-cloud-firestore==2.14.0
google-cloud-core==2.3.3
google-api-core==2.17.1

gunicorn==21.2.0

xgboost==2.0.3
numpy==1.26.4
pandas==2.2.0